backend/requirements.txt
bidict==0.23.1
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
Flask-Cors==5.0.0
flask-mongoengine==1.0.0
Flask-SocketIO==5.5.1
Flask-WTF==1.2.2
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mongoengine==0.29.1
pymongo==4.11.1
pyngrok==7.2.3
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
PyYAML==6.0.2
requests==2.32.3
setuptools==75.8.0
simple-websocket==1.1.0
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.45.1
wsproto==1.2.0
WTForms==3.2.1
